# swagger_output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EGI Media AI Backend API",
    "version": "0.1.0",
    "description": "API contract for the EGI Media AI decision-support service. The service is tenant- and company-scoped, reads editorial articles from egi-media-backend, and exposes only validated AI projections. It never exposes prompts, provider credentials, raw provider output, or CMS write operations."
  },
  "servers": [
    {
      "url": "http://localhost:5003",
      "description": "Local AI backend"
    }
  ],
...
  "paths": {
    "/api/v1/dashboard/executive-summary": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get the ranked Executive Summary",
        "description": "Returns at most five active, validated issues for one authorized company. The Top 5 cap exists only in this projection; it is not a storage, search, report, or analysis cap.",
        "operationId": "getExecutiveSummary",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompanyIdQuery" },
          { "$ref": "#/components/parameters/PeriodQuery" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "responses": {
          "200": { "description": "Validated Executive Summary", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExecutiveSummaryResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/BusinessRuleError" }
        }
      }
    },
    "/api/v1/issues": {
      "get": {
        "tags": ["Issues"],
        "summary": "Search and browse issues",
        "description": "Returns company-scoped issue projections. Results are not limited to Top 5. Only current validated data is returned; pending or failed stages are represented by safe pipeline state, not raw model output.",
        "operationId": "listIssues",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompanyIdQuery" },
          { "$ref": "#/components/parameters/PeriodQuery" },
          { "$ref": "#/components/parameters/PriorityQuery" },
          { "$ref": "#/components/parameters/IssueStatusQuery" },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/PageQuery" },
          { "$ref": "#/components/parameters/LimitQuery" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "responses": {
          "200": { "description": "Paginated issues", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IssueListResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/issues/{issueId}": {
      "get": {
        "tags": ["Issues"],
        "summary": "Get issue detail",
        "description": "Returns the issue timeline, current validated analysis, priority, claims, and canonical article citations. The caller must be authorized for the requested company.",
        "operationId": "getIssue",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/IssueIdPath" },
          { "$ref": "#/components/parameters/CompanyIdQuery" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "responses": {
          "200": { "description": "Issue detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IssueDetailResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/v1/companies/{companyId}/context": {
      "get": {
        "tags": ["Company Context"],
        "summary": "Get effective Company Context",
        "operationId": "getCompanyContext",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompanyIdPath" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "responses": {
          "200": { "description": "Effective Company Context", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Company Context"],
        "summary": "Create a new Company Context version",
        "description": "Human edit becomes the effective source of truth. The command creates a new version and never overwrites prior context in place.",
        "operationId": "putCompanyContext",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompanyIdPath" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" },
          { "$ref": "#/components/parameters/IfMatchHeader" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextWriteRequest" } } } },
        "responses": {
          "200": { "description": "New context version created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/BusinessRuleError" }
        }
      }
    },
    "/api/v1/company-context/draft": {
      "post": {
        "tags": ["Company Context"],
        "summary": "Request an asynchronous Company Context draft",
        "description": "Starts T01. URL/file/manual text is untrusted. Backend validates source policy before a worker can call the model; the result is always a draft and never automatically effective.",
        "operationId": "requestCompanyContextDraft",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompanyIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" },
          { "$ref": "#/components/parameters/RequestIdHeader" }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextDraftRequest" } } } },
        "responses": {
          "202": { "description": "Draft pipeline run accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PipelineRunResponse" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/company-context/drafts/{draftId}": {
      "get": {
        "tags": ["Company Context"],
        "summary": "Get a Company Context draft for human review",
        "operationId": "getCompanyContextDraft",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/DraftIdPath" }, { "$ref": "#/components/parameters/RequestIdHeader" }],
        "responses": {
          "200": { "description": "Company Context draft", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextDraftResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Company Context"],
        "summary": "Edit a Company Context draft before approval",
        "description": "Human edits move an in-review draft back to draft. Approved drafts cannot be edited.",
        "operationId": "editCompanyContextDraft",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/DraftIdPath" }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextDraftEditRequest" } } } },
        "responses": {
          "200": { "description": "Edited draft", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextDraftResponse" } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/company-context/drafts/{draftId}/submit-review": {
      "post": {
        "tags": ["Company Context"],
        "summary": "Submit a Company Context draft for human review",
        "operationId": "submitCompanyContextDraftReview",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/DraftIdPath" }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }],
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextReviewRequest" } } } },
        "responses": {
          "200": { "description": "Draft is in review", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextDraftResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/company-context/drafts/{draftId}/approve": {
      "post": {
        "tags": ["Company Context"],
        "summary": "Approve a reviewed Company Context draft",
        "description": "Only this human-approved action creates an effective Company Context version for downstream T02 use.",
        "operationId": "approveCompanyContextDraft",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/DraftIdPath" }, { "$ref": "#/components/parameters/IdempotencyKeyHeader" }, { "$ref": "#/components/parameters/RequestIdHeader" }],
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextReviewRequest" } } } },
        "responses": {
          "200": { "description": "Draft approved and effective context version created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyContextDraftApprovalResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "End-user identity. Tenant/company scope is derived and verified server-side." },
      "internalServiceAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "Service JWT", "description": "Internal service or authorized operator identity; never available to frontend." }
    },
    "parameters": {
      "CompanyIdQuery": { "name": "company_id", "in": "query", "required": true, "description": "Company context to read. Server verifies tenant membership.", "schema": { "$ref": "#/components/schemas/Uuid" } },
      "CompanyIdHeader": { "name": "X-Company-Id", "in": "header", "required": true, "description": "Active company for a mutation. Server verifies tenant/company access.", "schema": { "$ref": "#/components/schemas/Uuid" } },
      "PeriodQuery": { "name": "period", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/Period" } },
      "PriorityQuery": { "name": "priority", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/Priority" } },
      "IssueStatusQuery": { "name": "status", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/IssueStatus" } },
...
      "SearchQuery": { "name": "q", "in": "query", "required": false, "schema": { "type": "string", "minLength": 1, "maxLength": 200 } },
      "PageQuery": { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "LimitQuery": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
      "IssueIdPath": { "name": "issueId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
      "CompanyIdPath": { "name": "companyId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
      "DraftIdPath": { "name": "draftId", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 255 } },
...
      "IdempotencyKeyHeader": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 16, "maxLength": 255 } },
      "IfMatchHeader": { "name": "If-Match", "in": "header", "required": true, "description": "Current resource version for optimistic concurrency.", "schema": { "type": "string", "pattern": "^[0-9]+$" } },
      "RequestIdHeader": { "name": "X-Request-Id", "in": "header", "required": false, "schema": { "type": "string", "maxLength": 100 } }
    },
    "responses": {
      "Unauthorized": { "description": "Authentication missing or invalid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Forbidden": { "description": "Caller lacks required tenant/company/role permission", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "NotFound": { "description": "Resource not found in caller scope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "Conflict": { "description": "Idempotency key, version, or state conflict", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "ValidationError": { "description": "Request schema validation failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
      "BusinessRuleError": { "description": "Valid request cannot proceed because of pipeline/state/delivery business rules", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
...
    },
    "schemas": {
      "Uuid": { "type": "string", "format": "uuid" },
      "Timestamp": { "type": "string", "format": "date-time" },
      "Period": { "type": "string", "enum": ["24jam", "7hari", "30hari"], "default": "24jam" },
...
      "Priority": { "type": "string", "enum": ["tinggi", "sedang", "rendah"] },
      "IssueStatus": { "type": "string", "enum": ["baru", "berkembang", "dipantau", "selesai"] },
...
      "PipelineStageState": { "type": "string", "enum": ["queued", "running", "succeeded", "failed", "retrying", "dead_letter", "waiting_review", "insufficient_data", "stale", "cancelled"] },
      "ErrorCode": { "type": "string", "enum": ["VALIDATION_ERROR", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "VERSION_CONFLICT", "IDEMPOTENCY_CONFLICT", "BUSINESS_RULE_FAILED", "AI_OUTPUT_INVALID", "SOURCE_VALIDATION_FAILED", "PIPELINE_PENDING", "INTERNAL_ERROR"] },
      "ApiMeta": {
        "type": "object",
        "required": ["request_id"],
        "properties": {
          "request_id": { "type": "string" },
          "trace_id": { "type": "string" },
          "page": { "type": "integer", "minimum": 1 },
          "limit": { "type": "integer", "minimum": 1 },
          "total": { "type": "integer", "minimum": 0 },
          "total_pages": { "type": "integer", "minimum": 0 }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error", "meta"],
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "type": "object", "required": ["code", "message"], "properties": { "code": { "$ref": "#/components/schemas/ErrorCode" }, "message": { "type": "string" }, "fields": { "type": "array", "items": { "type": "object", "properties": { "field": { "type": "string" }, "code": { "type": "string" } } } } } },
          "meta": { "$ref": "#/components/schemas/ApiMeta" }
        }
      },
...
      "IssueCard": {
        "type": "object",
        "required": ["id", "company_id", "title", "one_liner", "status", "priority", "last_developed_at", "pipeline"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "company_id": { "$ref": "#/components/schemas/Uuid" },
          "title": { "type": "string", "maxLength": 255 },
          "one_liner": { "type": "string", "maxLength": 500 },
          "status": { "$ref": "#/components/schemas/IssueStatus" },
          "priority": { "$ref": "#/components/schemas/Priority" },
          "priority_reason": { "type": "string", "maxLength": 1000 },
          "last_developed_at": { "$ref": "#/components/schemas/Timestamp" },
          "is_saved": { "type": "boolean" },
          "pipeline": { "type": "array", "items": { "$ref": "#/components/schemas/PipelineStage" } }
        }
      },
      "IssueDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/IssueCard" },
          { "type": "object", "required": ["developments", "citations"], "properties": {
            "developments": { "type": "array", "items": { "$ref": "#/components/schemas/IssueDevelopment" } },
            "analysis": { "$ref": "#/components/schemas/IssueAnalysis" },
            "citations": { "type": "array", "items": { "$ref": "#/components/schemas/SourceCitation" } },
            "version": { "type": "integer", "minimum": 1 }
          } }
        ]
      },
...
      "ExecutiveSummary": {
        "type": "object",
        "required": ["company_id", "period", "generated_at", "issues"],
        "properties": {
          "company_id": { "$ref": "#/components/schemas/Uuid" },
          "period": { "$ref": "#/components/schemas/Period" },
          "generated_at": { "$ref": "#/components/schemas/Timestamp" },
          "issues": { "type": "array", "maxItems": 5, "items": { "$ref": "#/components/schemas/IssueCard" } }
        }
      },
...
      "CompanyContext": {
        "type": "object",
        "required": ["company_id", "version", "status", "source", "updated_at", "fields"],
        "properties": {
          "company_id": { "$ref": "#/components/schemas/Uuid" },
          "version": { "type": "integer", "minimum": 1 },
          "status": { "type": "string", "enum": ["draft", "effective", "archived"] },
          "source": { "type": "string", "enum": ["manual", "ai_draft", "import"] },
          "updated_at": { "$ref": "#/components/schemas/Timestamp" },
          "updated_by": { "type": "string", "maxLength": 255 },
          "fields": { "$ref": "#/components/schemas/CompanyContextFields" }
        }
      },
      "CompanyContextFields": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": { "type": ["string", "null"], "maxLength": 255 }, "industry": { "type": ["string", "null"], "maxLength": 255 }, "sub_industry": { "type": ["string", "null"], "maxLength": 255 }, "description": { "type": ["string", "null"], "maxLength": 4000 },
          "products": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "customers": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "regions": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "competitors": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } },
          "priorities": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "goals": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "risks": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "topics": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "dependencies": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }
        }
      },
      "CompanyContextWriteRequest": { "type": "object", "required": ["version", "fields"], "additionalProperties": false, "properties": { "version": { "type": "integer", "minimum": 1 }, "fields": { "$ref": "#/components/schemas/CompanyContextFields" }, "change_reason": { "type": "string", "maxLength": 1000 } } },
...
      "CompanyContextDraftEditRequest": { "type": "object", "required": ["fields"], "additionalProperties": false, "properties": { "fields": { "type": "object", "additionalProperties": false, "properties": { "name": { "type": ["string", "null"], "maxLength": 255 }, "industry": { "type": ["string", "null"], "maxLength": 255 }, "sub_industry": { "type": ["string", "null"], "maxLength": 255 }, "description": { "type": ["string", "null"], "maxLength": 4000 }, "products": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "customers": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "regions": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "competitors": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "priorities": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "goals": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "risks": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "topics": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } }, "dependencies": { "type": "array", "maxItems": 30, "items": { "type": "string", "maxLength": 255 } } } }, "review_note": { "type": "string", "maxLength": 1000 } } },
      "CompanyContextReviewRequest": { "type": "object", "additionalProperties": false, "properties": { "review_note": { "type": "string", "maxLength": 1000 }, "approval_note": { "type": "string", "maxLength": 1000 } } },
      "CompanyContextDraftRequest": { "type": "object", "required": ["source"], "additionalProperties": false, "properties": { "source": { "oneOf": [{ "$ref": "#/components/schemas/ContextUrlSource" }, { "$ref": "#/components/schemas/ContextTextSource" }] } } },
      "ContextUrlSource": { "type": "object", "required": ["type", "url"], "additionalProperties": false, "properties": { "type": { "type": "string", "enum": ["url"] }, "url": { "type": "string", "format": "uri", "maxLength": 2048 } } },
      "ContextTextSource": { "type": "object", "required": ["type", "text"], "additionalProperties": false, "properties": { "type": { "type": "string", "enum": ["text"] }, "text": { "type": "string", "minLength": 1, "maxLength": 100000 } } },
...
      "PipelineRun": { "type": "object", "required": ["id", "trigger", "state", "stages", "created_at", "trace_id"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "trigger": { "type": "string", "enum": ["scheduled", "manual", "backfill", "replay", "context_draft"] }, "state": { "$ref": "#/components/schemas/PipelineStageState" }, "tenant_id": { "$ref": "#/components/schemas/Uuid" }, "company_id": { "$ref": "#/components/schemas/Uuid" }, "stages": { "type": "array", "items": { "$ref": "#/components/schemas/PipelineStage" } }, "created_at": { "$ref": "#/components/schemas/Timestamp" }, "completed_at": { "$ref": "#/components/schemas/Timestamp" }, "trace_id": { "type": "string" } } },
...
      "ExecutiveSummaryResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "$ref": "#/components/schemas/ExecutiveSummary" }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
      "IssueListResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/IssueCard" } } } }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
      "IssueDetailResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "$ref": "#/components/schemas/IssueDetail" }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
...
      "CompanyContextResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "$ref": "#/components/schemas/CompanyContext" }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
      "CompanyContextDraftResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "$ref": "#/components/schemas/CompanyContextDraft" }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
      "CompanyContextDraftApprovalResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "type": "object", "required": ["draft", "effective_context"], "properties": { "draft": { "$ref": "#/components/schemas/CompanyContextDraft" }, "effective_context": { "$ref": "#/components/schemas/CompanyContext" } } }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
      "PipelineRunResponse": { "type": "object", "required": ["success", "data", "meta"], "properties": { "success": { "type": "boolean", "enum": [true] }, "data": { "$ref": "#/components/schemas/PipelineRun" }, "meta": { "$ref": "#/components/schemas/ApiMeta" } } },
...
    }
  }
}